# openapi.yaml
openapi: 3.0.3
info:
  title: CatWAF API
  version: 1.0.1
  description: |
    CatWAF's control-plane API — everything the dashboard itself calls.
    There's no separate "public API" here; this is the real thing the
    frontend uses, documented so you can script against it or build your
    own tooling on top.

    **Scope of this document:** CatWAF has ~167 routes across 19 files.
    This spec covers the core groups in real depth — auth and two-factor
    login, WAF engine settings, custom rules, IP/geo lists, the security
    score, honeypots, sensitive-file protection, operational modes
    (panic/maintenance/config signing), and the settings, protection and
    operations layers — rather than a shallow stub for every endpoint.
    Routes not yet documented here still exist and work; see
    `backend/routes/` for the full list. Expanding this file's coverage is
    tracked in `ROADMAP.md`.

    **The settings layer is self-describing.** Rather than enumerating ~290
    fields here, `GET /api/settings/schema` returns the live schema the
    backend validates against and the dashboard builds its controls from —
    field names, types, defaults, permitted values and help text. That is
    the authoritative reference; `docs/settings.md` is generated from it.

    **One thing worth knowing:** the dashboard stats/traffic endpoints
    (`/api/stats`, `/api/traffic/*`) return real data — sourced from
    Coraza's own audit log via `services/requestLog.js`, not a demo feed —
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Obtained from POST /api/auth/login. Unauthenticated requests are
        still accepted for GET routes (treated as the built-in `guest`/
        viewer role) — the token only matters for anything that mutates
        state, which requires the `admin` role.
...
paths:
...
  /api/jobs:
    get:
      tags: [Operations]
      summary: Every scheduled job, its interval and its last run
      description: A job whose feature is switched off is listed but never scheduled.
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/jobs/{name}/run:
    post:
      tags: [Operations]
      summary: Run one job now, regardless of its schedule
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Finished }
        '404': { description: No such job }
        '500': { description: The job ran and failed }
  /api/caches:
    get:
      tags: [Operations]
      summary: Cache namespaces, entry counts and sizes
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/caches/{id}/clear:
    post:
      tags: [Operations]
      summary: Empty one namespace ("all" empties every namespace)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Cleared }
  /api/templates:
    get:
      tags: [Operations]
      summary: Built-in and saved configuration templates
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
    post:
      tags: [Operations]
      summary: Save the current configuration as a template
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string }, groups: { type: array, items: { type: string }, nullable: true }, include_waf: { type: boolean, default: true }, description: { type: string } } } } }
      responses:
        '200': { description: Saved }
  /api/templates/{id}/apply:
    post:
      tags: [Operations]
      summary: Apply a template, or show what it would change
      description: >
        Send dry_run true first. Applying rewrites several settings groups at
        once, which is exactly the change worth reading before it happens.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content: { application/json: { schema: { type: object, properties: { dry_run: { type: boolean, default: false } } } } }
      responses:
        '200': { description: 'Applied, or the change list when dry_run is true' }
  /api/reports:
    get:
      tags: [Operations]
      summary: Activity over a date range
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: format, in: query, schema: { type: string, enum: [json, csv, events-csv, html], default: json } }
      responses:
        '200':
          description: OK
          content:
            application/json: { schema: { type: object } }
            text/csv: { schema: { type: string } }
            text/html: { schema: { type: string } }
  /api/backups:
    get:
      tags: [Operations]
      summary: Stored backups and the configured destination
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/backups/run:
    post:
      tags: [Operations]
      summary: Write a backup now
      security: [{ bearerAuth: [] }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { dry_run: { type: boolean }, destination: { type: string, nullable: true } } } } }
      responses:
        '200': { description: OK }
  /api/telemetry:
    get:
      tags: [Operations]
      summary: Telemetry status and the exact payload that would be sent
      description: >
        payload_preview is the real payload built from this install right now,
        not an illustration of one. never_collected states what is excluded.
        Telemetry is off unless explicitly enabled.
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
...
  /api/tls/status:
    get:
      tags: [TLS]
      summary: Certificates Caddy holds, ACME advice, and Caddy's capabilities
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/tls/validate:
    post:
      tags: [TLS]
      summary: Inspect a certificate, or check a certificate and key match
      description: >
        A mismatched certificate and key does not fail at apply time — it fails
        when a visitor tries to connect. Check first.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, properties: { certificate: { type: string }, key: { type: string }, ca: { type: string } } } } }
      responses:
        '200': { description: 'Inspection result; ok is false with an error when the certificate is unusable' }

  # ─── Two-factor login ──────────────────────────────────────────────────
  /api/auth/2fa:
    get:
      tags: [Auth]
      summary: Whether two-factor is enabled for the current account
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/auth/2fa/enroll:
    post:
      tags: [Auth]
      summary: Begin enrollment
      description: >
        Returns the secret and an otpauth:// URI. Nothing is enforced until
        /confirm proves the authenticator app works — enrolling without that
        check is how people lock themselves out.
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
        '400': { description: Already enabled }
  /api/auth/2fa/confirm:
    post:
      tags: [Auth]
      summary: Prove the authenticator works and turn two-factor on
      description: >
        Returns ten recovery codes, once. They are stored as HMACs, so this
        response cannot be reproduced later. The code used here is spent by
        the confirmation and cannot then be used to log in.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [code], properties: { code: { type: string } } } } }
      responses:
        '200': { description: Enabled }
        '400': { description: Wrong or expired code }
  /api/auth/2fa/disable:
    post:
      tags: [Auth]
      summary: Turn two-factor off
      description: Re-checks the password — disabling a second factor is exactly what a stolen session would attempt.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [password], properties: { password: { type: string } } } } }
      responses:
        '200': { description: Disabled }
        '401': { description: Password confirmation failed }

  # ─── Plugins ───────────────────────────────────────────────────────────
  /api/plugins:
    get:
      tags: [Plugins]
      summary: Installed plugins, and the policy they are held to
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
    post:
      tags: [Plugins]
      summary: Install a plugin manifest
      description: >
        Plugins are data only. A manifest declaring code, script, main,
        require, exec, command, hooks, middleware, entry or eval is refused
        rather than ignored, so a plugin written against a nonexistent
        execution API fails loudly instead of appearing to work.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [catwaf_plugin, id], additionalProperties: true } } }
      responses:
        '200': { description: Installed }
        '400': { description: The manifest was refused; detail says why }